+++ app/src/main/java/com/gnz/getamovie/application/BuildersModule.kt
@@ -1,7 +1,7 @@
 package com.gnz.getamovie.application
 
 import com.gnz.getamovie.application.scopes.FragmentScope
-import com.gnz.getamovie.features.nowplaying.NowPlayingFragment
+import com.gnz.getamovie.features.movielist.MovieListFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -11,5 +11,5 @@ abstract class BuildersModule {
 
     @FragmentScope
     @ContributesAndroidInjector
-    abstract fun bindNowPlayingFragment(): NowPlayingFragment
+    abstract fun bindNowPlayingFragment(): MovieListFragment
 }
