+++ app/src/main/java/com/gnz/getamovie/features/MainActivity.kt
@@ -1,15 +1,11 @@
 package com.gnz.getamovie.features
 
-import android.app.SearchManager
-import android.content.Context
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.support.v7.widget.SearchView
-import android.view.Menu
 import com.gnz.getamovie.R
 import com.gnz.getamovie.features.moviedetails.MovieDetailsFragment
-import com.gnz.getamovie.features.nowplaying.NowPlayingFragment
-import com.gnz.getamovie.features.nowplaying.pagination.MovieDetails
+import com.gnz.getamovie.features.movielist.MovieListFragment
+import com.gnz.getamovie.features.movielist.pagination.MovieDetails
 
 class MainActivity : AppCompatActivity() {
 
@@ -19,7 +15,7 @@ class MainActivity : AppCompatActivity() {
 
         supportFragmentManager.beginTransaction()
                 .setCustomAnimations(android.R.animator.fade_in, android.R.animator.fade_out)
-                .replace(R.id.containerLayout, NowPlayingFragment.newInstance(), "MOVIE")
+                .replace(R.id.containerLayout, MovieListFragment.newInstance(), "MOVIE")
                 .commit()
     }
