+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".features.nowplaying.NowPlayingFragment">
+    tools:context=".features.movielist.MovieListFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/nowPlayingRecyclerView"
